blend_api/pyproject.toml
[project]
name = "blend_api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "functions-framework>=3.8.2",
    "lkr-dev-cli>=0.0.32",
    "looker-sdk>=25.4.0",
    "pydantic>=2.10.6",
    "pygithub>=2.6.1",
    "structlog>=25.2.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
]
